move imagecollector controller to separate folder
